+++ back/controllers/upload/getStorePath.js
@@ -2,7 +2,7 @@
 
 const fs = require('fs').promises;
 const path = require('path');
-const { fileExists } = require('./fileUpload-utils');
+const { fileExists } = require('../utils/utils-controller');
 
 /**
  * Returns the path to store the pictures, if if not exists creates a new one
